requirements.txt
aiogram
aiohttp

yt-dlp
shortuuid

psycopg2-binary
asyncpg
aioredis

sqlalchemy
python-dotenv

